+++ CHANGELOG.md
@@ -10,6 +10,7 @@ project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 Changed `DOMPoint()` constructor to check for parameter nullability.
 ### Added
+* Added `deregisterAllFonts` method to free up memory and reduce font conflicts.
 ### Fixed
 
 2.8.0